move question in its own file; remove styles logic
