+++ src/app/TakeTest/createANewTest.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 
 class CreateANewTest extends React.Component {
-  constructor(props, context) {
-    super(props, context);
+  constructor(props) {
+    super(props);
     this.state = {
       candidateName: '',
       testType: this.getTestTypes()[0]
